+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Moved holidays object from `moment.fn.holidays` to `moment.holidays`.
 - Fixed bug where `holidaysBetween` function would only return holidays for the start year.
 - `holidaysBetween` function now returns an array of moment objects rather than an object.
+- `isHoliday` now accepts a `holidays` parameter to get a `true` response.
 
 
 ## 1.0.0 - 2017-07-16